Added AccessControlService to restrict user access by roles to ToDoController as test

+++ src/main/java/com/example/Todo_list/controller/ToDoController.java
@@ -3,6 +3,7 @@ package com.example.Todo_list.controller;
 import com.example.Todo_list.entity.Task;
 import com.example.Todo_list.entity.ToDo;
 import com.example.Todo_list.entity.User;
+import com.example.Todo_list.security.AccessControlService;
 import com.example.Todo_list.service.NotificationService;
 import com.example.Todo_list.service.TaskService;
 import com.example.Todo_list.service.ToDoService;
@@ -11,13 +12,16 @@ import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
 import java.time.ZonedDateTime;
+import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.List;
 import java.util.Objects;
@@ -35,6 +39,7 @@ public class ToDoController {
     private final TaskService taskService;
     private final UserService userService;
     private final NotificationService notificationService;
+    private final AccessControlService accessControlService;
 
     /**
      * Display the form for creating a new ToDo item
@@ -44,7 +49,12 @@ public class ToDoController {
      */
     @PreAuthorize("hasAuthority('ADMIN') or #ownerId == authentication.principal.id")
     @GetMapping("/create/users/{owner_id}")
-    public String showToDoCreationForm(@PathVariable("owner_id") Long ownerId, Model model) {
+    public String showToDoCreationForm(Authentication authentication, @PathVariable("owner_id") Long ownerId, Model model) {
+        if (!accessControlService.hasAccess(authentication, ownerId)) {
+            logger.info("ToDoController.showToDoCreationForm(): User does not have access to create ToDo");
+            throw new AccessDeniedException("You do not have access to create a Project for this user");
+        }
+
         model.addAttribute("todo", new ToDo());
         model.addAttribute("ownerId", ownerId);
         logger.info("ToDoController.showToDoCreationForm(): Displaying ToDo creation form");
@@ -61,10 +71,16 @@ public class ToDoController {
     @PreAuthorize("hasAuthority('ADMIN') or #ownerId == authentication.principal.id")
     @PostMapping("/create/users/{owner_id}")
     public String createToDo(
+            Authentication authentication,
             @PathVariable("owner_id") Long ownerId,
             @Valid @ModelAttribute("todo") ToDo toDo,
             BindingResult result
     ) {
+        if (!accessControlService.hasAccess(authentication, ownerId)) {
+            logger.info("ToDoController.createToDo(): User does not have access to create ToDo");
+            throw new AccessDeniedException("You do not have access to create a Project for this user");
+        }
+
         logger.info("ToDoController.createToDo(): Attempting to create a ToDo item for user with userId=" + ownerId);
 
         if (result.hasErrors()) {
@@ -90,8 +106,16 @@ public class ToDoController {
             "principal.id == @toDoServiceImpl.findToDoById(#todoId).owner.id or " +
             "@toDoServiceImpl.findToDoById(#todoId).collaborators.contains(@userServiceImpl.findUserById(principal.id))")
     @GetMapping("/{todo_id}/tasks")
-    public String displayToDo(@PathVariable("todo_id") Long todoId, Model model) {
+    public String displayToDo(Authentication authentication, @PathVariable("todo_id") Long todoId, Model model) {
         ToDo todo = toDoService.findToDoById(todoId);
+        List<Long> permittedUserIds = new ArrayList<>(List.of(todo.getOwner().getId()));
+        permittedUserIds.addAll(todo.getCollaborators().stream().map(User::getId).toList());
+
+        if (!accessControlService.hasAccess(authentication, permittedUserIds)) {
+            logger.info("ToDoController.displayToDo(): User does not have access to view ToDo");
+            throw new AccessDeniedException("You do not have access to view this Project");
+        }
+
         List<Task> tasks = taskService.findAllTasksOfToDo(todoId);
         List<User> users = userService.findAllUsers()
                                         .stream()
@@ -115,12 +139,25 @@ public class ToDoController {
      * @param model the model to be used in the view
      * @return the view to be displayed
      */
+    @PreAuthorize("hasAuthority('ADMIN') or " +
+            "principal.id == @toDoServiceImpl.findToDoById(#todoId).owner.id or " +
+            "@toDoServiceImpl.findToDoById(#todoId).collaborators.contains(@userServiceImpl.findUserById(principal.id))")
     @GetMapping("/{todo_id}/update/users/{owner_id}")
     public String showToDoUpdateForm(
+            Authentication authentication,
             @PathVariable("todo_id") Long todoId,
             @PathVariable("owner_id") Long ownerId,
             Model model
     ) {
+        ToDo todo = toDoService.findToDoById(todoId);
+        List<Long> permittedUserIds = new ArrayList<>(List.of(todo.getOwner().getId()));
+        permittedUserIds.addAll(todo.getCollaborators().stream().map(User::getId).toList());
+
+        if (!accessControlService.hasAccess(authentication, permittedUserIds)) {
+            logger.info("ToDoController.showToDoUpdateForm(): User does not have access to update ToDo");
+            throw new AccessDeniedException("You do not have access to update this Project");
+        }
+
         ToDo toDo = toDoService.findToDoById(todoId);
         model.addAttribute("todo", toDo);
         logger.info("ToDoController.showToDoUpdateForm(): Displaying update form of " + toDo);
@@ -140,11 +177,21 @@ public class ToDoController {
                 "@toDoServiceImpl.findToDoById(#todoId).collaborators.contains(@userServiceImpl.findUserById(principal.id))")
     @PostMapping("/{todo_id}/update/users/{owner_id}")
     public String updateToDo(
+            Authentication authentication,
             @PathVariable("todo_id") Long todoId,
             @PathVariable("owner_id") Long ownerId,
             @Valid @ModelAttribute("todo") ToDo todo,
             BindingResult result)
     {
+        ToDo oldTodo = toDoService.findToDoById(todoId);
+        List<Long> permittedUserIds = new ArrayList<>(List.of(oldTodo.getOwner().getId()));
+        permittedUserIds.addAll(todo.getCollaborators().stream().map(User::getId).toList());
+
+        if (!accessControlService.hasAccess(authentication, permittedUserIds)) {
+            logger.info("ToDoController.updateToDo(): User does not have access to update ToDo");
+            throw new AccessDeniedException("You do not have access to update this Project");
+        }
+
         logger.info("ToDoController.updateToDo(): Attempting to update ToDo with toDoId=" + todoId);
 
         if (result.hasErrors()) {
@@ -153,7 +200,6 @@ public class ToDoController {
             return "todo-update";
         }
 
-        ToDo oldTodo = toDoService.findToDoById(todoId);
         todo.setOwner(oldTodo.getOwner());
         todo.setCollaborators(oldTodo.getCollaborators());
 
@@ -172,7 +218,17 @@ public class ToDoController {
                 "principal.id == @toDoServiceImpl.findToDoById(#todoId).owner.id or " +
                 "@toDoServiceImpl.findToDoById(#todoId).collaborators.contains(@userServiceImpl.findUserById(principal.id))")
     @PostMapping("/{todo_id}/delete/users/{owner_id}")
-    public String deleteToDo(@PathVariable("todo_id") Long todoId, @PathVariable("owner_id") Long ownerId) {
+    public String deleteToDo(Authentication authentication, @PathVariable("todo_id") Long todoId,
+                             @PathVariable("owner_id") Long ownerId) {
+        ToDo todo = toDoService.findToDoById(todoId);
+        List<Long> permittedUserIds = new ArrayList<>(List.of(todo.getOwner().getId()));
+        permittedUserIds.addAll(todo.getCollaborators().stream().map(User::getId).toList());
+
+        if (!accessControlService.hasAccess(authentication, permittedUserIds)) {
+            logger.info("ToDoController.deleteToDo(): User does not have access to delete ToDo");
+            throw new AccessDeniedException("You do not have access to delete this Project");
+        }
+
         logger.info("ToDoController.deleteToDo(): Deleting ToDo with toDoId=" + todoId);
         toDoService.deleteToDoById(todoId);
         return String.format("redirect:/todos/all/users/%d", ownerId);
@@ -186,7 +242,12 @@ public class ToDoController {
      */
     @PreAuthorize("hasAuthority('ADMIN') or #userId == authentication.principal.id")
     @GetMapping("/all/users/{user_id}")
-    public String displayAllToDosOfUser(@PathVariable("user_id") Long userId, Model model) {
+    public String displayAllToDosOfUser(Authentication authentication, @PathVariable("user_id") Long userId, Model model) {
+        if (!accessControlService.hasAccess(authentication, userId)) {
+            logger.info("ToDoController.displayAllToDo(): User does not have access to view ToDos");
+            throw new AccessDeniedException("You do not have access to view Projects for this user");
+        }
+
         List<ToDo> todos = toDoService.findAllToDoOfUserId(userId);
         List<ToDo> sortedTodos = todos.stream().sorted(Comparator.comparingLong(ToDo::getId)).toList();
         model.addAttribute("todos", sortedTodos);
@@ -203,7 +264,12 @@ public class ToDoController {
      */
     @PreAuthorize("hasAuthority('ADMIN')")
     @GetMapping("/all")
-    public String displayAllToDos(@RequestParam("user_id") Long userId, Model model) {
+    public String displayAllToDos(Authentication authentication, @RequestParam("user_id") Long userId, Model model) {
+        if (!accessControlService.hasAdminAccess(authentication)) {
+            logger.info("ToDoController.displayAllToDo(): User does not have access to view all ToDos");
+            throw new AccessDeniedException("You do not have access to view all Projects");
+        }
+
         List<ToDo> todos = toDoService.findAllToDos();
         List<ToDo> sortedTodos = todos.stream().sorted(Comparator.comparingLong(ToDo::getId)).toList();
         model.addAttribute("todos", sortedTodos);
@@ -220,7 +286,13 @@ public class ToDoController {
      */
     @PreAuthorize("hasAuthority('ADMIN') or authentication.principal.id == @toDoServiceImpl.findToDoById(#todoId).owner.id")
     @PostMapping("/{todo_id}/add")
-    public String addCollaborator(@PathVariable("todo_id") Long todoId, @RequestParam("user_id") Long userId) {
+    public String addCollaborator(Authentication authentication, @PathVariable("todo_id") Long todoId,
+                                  @RequestParam("user_id") Long userId) {
+        if (!accessControlService.hasAccess(authentication, toDoService.findToDoById(todoId).getOwner().getId())) {
+            logger.info("ToDoController.addCollaborator(): User does not have access to add collaborator");
+            throw new AccessDeniedException("You do not have access to add a collaborator to this Project");
+        }
+
         if (userId == -1) {
             logger.info("ToDoController.addCollaborator(): userId is invalid: " + userId);
             return String.format("redirect:/todos/%d/tasks", todoId);
@@ -249,7 +321,13 @@ public class ToDoController {
      */
     @PreAuthorize("hasAuthority('ADMIN') or authentication.principal.id == @toDoServiceImpl.findToDoById(#todoId).owner.id")
     @PostMapping("/{todo_id}/remove")
-    public String removeCollaborator(@PathVariable("todo_id") Long todoId, @RequestParam("user_id") Long userId) {
+    public String removeCollaborator(Authentication authentication, @PathVariable("todo_id") Long todoId,
+                                     @RequestParam("user_id") Long userId) {
+        if (!accessControlService.hasAccess(authentication, toDoService.findToDoById(todoId).getOwner().getId())) {
+            logger.info("ToDoController.addCollaborator(): User does not have access to add collaborator");
+            throw new AccessDeniedException("You do not have access to add a collaborator to this Project");
+        }
+
         ToDo todo = toDoService.findToDoById(todoId);
         logger.info("ToDoController.removeCollaborator(): " +
                     "Removing " + todo + " from " + userService.findUserById(userId));
